[Feat] 리스트 컴포넌트 분리/바텀시트 수정
